% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ames_missing}
\alias{ames_missing}
\title{Ames Housing Data with explicit missings}
\source{
De Cock, D. (2011). "Ames, Iowa: Alternative to the Boston Housing Data as an End of Semester Regression Project," \emph{Journal of Statistics Education},  Volume 19, Number 3.

\url{http://jse.amstat.org/v19n3/decock/DataDocumentation.txt}

\url{http://jse.amstat.org/v19n3/decock.pdf}
}
\value{
\item{ames_missing}{a tibble}
}
\description{
A data set from De Cock (2011) has 82 fields were recorded for 2,930
properties in Ames IA. This version is copies from the \code{AmesHousing} package
but
\itemize{
\item does not include a few quality columns that appear to be outcomes
rather than predictors.
\item has missing values explicitely encoded as missing data like for example
\code{Garage_Area} is \code{NA} when \code{Garage_Cond}-ition is \code{No_Garage}
}
}
\details{
See this links for the sources below for more information as well as
\code{?AmesHousing::make_ames}.
}
\examples{
data(ames_missing)
str(ames_missing)
}
\keyword{datasets}
