% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.tabnet_fit}
\alias{autoplot.tabnet_fit}
\title{Plot tabnet_fit loss along epochs}
\usage{
autoplot.tabnet_fit(object, ...)
}
\arguments{
\item{object}{A \code{tabnet_fit} object as a result of \code{\link[=tabnet_fit]{tabnet_fit()}} .}

\item{...}{For plots with a regular grid, this is passed to \code{format()} and is
applied to a parameter used to color points. Otherwise, it is not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot tabnet_fit loss along epochs
}
\details{

}
\examples{
\donttest{
data("ames", package = "modeldata")
ames_fit <- tabnet_fit(Sale_Price ~. , data=ames, epoch=15)

# Plot the model loss over epochs
autoplot(ames_fit)
}
}
