% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.tabnet_pretrain}
\alias{autoplot.tabnet_pretrain}
\title{Plot tabnet_pretrain model loss along epochs}
\usage{
\method{autoplot}{tabnet_pretrain}(object, ...)
}
\description{
Plot tabnet_pretrain model loss along epochs
}
