% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grouped_feature.R
\name{create_group_matrix}
\alias{create_group_matrix}
\title{Create the group matrix corresponding to the given vars_id_groups}
\usage{
create_group_matrix(vars_id_groups, input_dim)
}
\arguments{
\item{vars_id_groups}{: list of vars_id_group, vars_id_group being the vector of features ids to group
Each element represents features in the same group. Feature id cannot be present in
multiple group, ids that are not part of a group will be implicitly part of a single feature group.}

\item{input_dim}{: number of feature in the initial dataset}
}
\value{
a torch_tensor of size (n_groups, input_dim)
where m_ij represents the importance of feature j in group i
The row sum must be 1 as each group is equally important a priori.
}
\description{
Create the group matrix corresponding to the given vars_id_groups
}
\examples{
\dontshow{if (torch::torch_is_installed()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# mtcars example
colnames(mtcars)
create_group_matrix(vars_id_groups = list(c(1,3), c(4:7), c(9:10)), input_dim = 11)
\dontshow{\}) # examplesIf}
}
