% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-type.R
\name{entmax15}
\alias{entmax15}
\title{1.5-entmax}
\usage{
entmax15(dim = -1L, k = NULL)
}
\arguments{
\item{dim}{The dimension along which to apply 1.5-entmax.}

\item{k}{The number of largest elements to partial-sort input over. For optimal
performance, should be slightly bigger than the expected number of
nonzeros in the solution. If the solution is more than k-sparse,
this function is recursively called with a 2*k schedule. If \code{NULL}, full
sorting is performed from the beginning.}

\item{input}{The input tensor.}
}
\value{
The projection result P  of the same shape as input, such that
\eqn{P$sum(dim=dim) == 1} elementwise.
}
\description{
Normalizing sparse transform (a la softmax).
}
\details{
Solves the optimization problem:
\eqn{\max_p <input, P> - H_{1.5}(P) \text{ s.t. } P \geq 0, \sum(P) == 1}
where \eqn{H_{1.5}(P)} is the Tsallis alpha-entropy with \eqn{\alpha=1.5}.
}
\examples{
# example usage
X <- torch::torch_randn(10,5)
dim <- 1
k <- 3
result <- entmax15(X, dim, k)
print(result)
}
