% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-network.R
\name{gbn}
\alias{gbn}
\title{Ghost Batch Normalization
https://arxiv.org/abs/1705.08741}
\usage{
gbn(input_dim, virtual_batch_size = 128, momentum = 0.01)
}
\description{
Ghost Batch Normalization
https://arxiv.org/abs/1705.08741
}
