% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{nn_maum_loss}
\alias{nn_maum_loss}
\title{Multiclass AUM loss}
\usage{
nn_maum_loss()
}
\description{
Creates a criterion that measures the Area under the \eqn{Min(FPR, FNR)} (AUM) between each
element in the input \eqn{pred_tensor} and target \eqn{label_tensor} for each minority class considered as a
1-versus-all binary classifier. All the minority class being all classes but the majority class.
}
\details{
This is used for measuring the error of a binary reconstruction within highly unbalanced dataset,
where the goal is optimizing the ROC curve. Note that the targets \eqn{label_tensor} should be the factor
levels of the binary outcome, i.e. with values  between \code{1L} and \code{nlevels(label)}.
}
\examples{
\dontshow{if ((torch::torch_is_installed() && require("modeldata"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(ames, package = "modeldata")
loss <- nn_maum_loss()
target <- torch::torch_tensor(ames$Bldg_Type)
input <- torch::torch_rand(target$shape, requires_grad = TRUE)
output <- loss(input, target)
output$backward()
\dontshow{\}) # examplesIf}
}
