% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{resolve_data}
\alias{resolve_data}
\title{Transforms input data into a list of
3 torch_tensors being
$x , $x_na_mask, $y
and
cat_idx the vector of x categorical predictor index
cat_dims the vector of number of levels of each x categorical predictor
input_dim and output_dim}
\usage{
resolve_data(x, y, device)
}
\arguments{
\item{x}{a data frame}

\item{y}{a response vector}

\item{device}{torch_device to send the tensor to}
}
\description{
Transforms input data into a list of
3 torch_tensors being
$x , $x_na_mask, $y
and
cat_idx the vector of x categorical predictor index
cat_dims the vector of number of levels of each x categorical predictor
input_dim and output_dim
}
