% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-type.R
\name{sparsemax}
\alias{sparsemax}
\title{Sparsemax}
\usage{
sparsemax(dim = -1L, k = NULL)
}
\arguments{
\item{dim}{The dimension along which to apply sparsemax.}

\item{k}{The number of largest elements to partial-sort input over. For optimal
performance, \code{k} should be slightly bigger than the expected number of
nonzeros in the solution. If the solution is more than k-sparse,
this function is recursively called with a 2*k schedule. If \code{NULL}, full
sorting is performed from the beginning.}

\item{input}{The input tensor.}
}
\value{
The projection result, such that \eqn{P$sum(dim=dim) ==1} elementwise.
}
\description{
Normalizing sparse transform (a la softmax).
}
\details{
Solves the projection:

\eqn{\min_P ||input - P||_2 \text{ s.t. } P \geq0, \sum(P) ==1}
}
\examples{
# example usage
X <- torch::torch.randn(10,5)
dim <-1
k <-3
result <- sparsemax(X, dim, k)
print(result)
}
