% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-network.R
\name{tabnet_no_embedding}
\alias{tabnet_no_embedding}
\title{Defines main part of the TabNet network without the embedding layers.}
\usage{
tabnet_no_embedding(
  input_dim,
  output_dim,
  n_d = 8,
  n_a = 8,
  n_steps = 3,
  gamma = 1.3,
  n_independent = 2,
  n_shared = 2,
  epsilon = 1e-15,
  virtual_batch_size = 128,
  momentum = 0.02,
  mask_type = "sparsemax"
)
}
\description{
Defines main part of the TabNet network without the embedding layers.
}
