% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials.R
\name{cat_emb_dim}
\alias{cat_emb_dim}
\alias{checkpoint_epochs}
\alias{drop_last}
\alias{encoder_activation}
\alias{lr_scheduler}
\alias{mlp_activation}
\alias{mlp_hidden_multiplier}
\alias{num_independent_decoder}
\alias{num_shared_decoder}
\alias{optimizer}
\alias{penalty}
\alias{verbose}
\alias{virtual_batch_size}
\title{Non-tunable parameters for the tabnet model}
\usage{
cat_emb_dim(range = NULL, trans = NULL)

checkpoint_epochs(range = NULL, trans = NULL)

drop_last(range = NULL, trans = NULL)

encoder_activation(range = NULL, trans = NULL)

lr_scheduler(range = NULL, trans = NULL)

mlp_activation(range = NULL, trans = NULL)

mlp_hidden_multiplier(range = NULL, trans = NULL)

num_independent_decoder(range = NULL, trans = NULL)

num_shared_decoder(range = NULL, trans = NULL)

optimizer(range = NULL, trans = NULL)

penalty(range = NULL, trans = NULL)

verbose(range = NULL, trans = NULL)

virtual_batch_size(range = NULL, trans = NULL)
}
\arguments{
\item{range}{unused}

\item{trans}{unused}
}
\description{
Non-tunable parameters for the tabnet model
}
