% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-vi-model.R
\name{vi_model}
\alias{vi_model}
\alias{vi_model.default}
\alias{vi_model.model_fit}
\alias{vi_model.workflow}
\title{Model-specific variable importance}
\source{
Johan Bring (1994) How to Standardize Regression Coefficients, The American
Statistician, 48:3, 209-213, DOI: 10.1080/00031305.1994.10476059.
}
\usage{
vi_model(object, ...)

\method{vi_model}{default}(object, ...)

\method{vi_model}{model_fit}(object, ...)

\method{vi_model}{workflow}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional optional arguments to be passed on to other methods.
See the details section below for arguments that can be passed to specific
object types.}

\item{type}{Character string specifying the type of variable importance to
return (only used for some models). See the details section below for which
methods this argument applies to.}

\item{lambda}{Numeric value for the penalty parameter of a model.}

\item{ncomp}{An integer for the number of partial least squares components
to be used in the importance calculations. If more components are requested
than were used in the model, all of the model's components are used.}
}
\value{
A tidy data frame (i.e., a \link[tibble:tibble]{tibble} object) with two
columns:
\itemize{
\item \code{Variable} - the corresponding feature name;
\item \code{Importance} - the associated importance, computed as the average change in
performance after a random permutation (or permutations, if \code{nsim > 1}) of
the feature in question.
}
}
\description{
Compute model-specific variable importance scores for the predictors in a
fitted model.
}
\details{
Computes model-specific variable importance scores depending on the class of
\code{object}:
}
\examples{
\dontrun{
# Basic example using imputed titanic data set
t3 <- titanic_mice[[1L]]

# Fit a simple model
set.seed(1449)  # for reproducibility
bst <- lightgbm::lightgbm(
  data = data.matrix(subset(t3, select = -survived)),
  label = ifelse(t3$survived == "yes", 1, 0),
  params = list("objective" = "binary", "force_row_wise" = TRUE),
  verbose = 0
)

# Compute VI scores
vi(bst)  # defaults to `method = "model"`
vi_model(bst)  # same as above

# Same as above (since default is `method = "model"`), but returns a plot
vip(bst, geom = "point")
vi_model(bst, type = "cover")
vi_model(bst, type = "cover", percentage = FALSE)

# Compare to
lightgbm::lgb.importance(bst)
}

}
