% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.tabnet_explain}
\alias{autoplot.tabnet_explain}
\title{Plot tabnet_explain interpretation}
\usage{
\method{autoplot}{tabnet_explain}(object, type = c("mask_agg", "steps"), ...)
}
\arguments{
\item{object}{A \code{tabnet_explain} object as a result of \code{\link[=tabnet_explain]{tabnet_explain()}}.}

\item{type}{a character value. Either \code{"mask_agg"} the default, for per predictor
mask importance aggregate,or \code{"steps"} for a plot at each mask step.}

\item{...}{not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot tabnet_explain interpretation
}
\details{
Plot the tabnet_explain object mask importance per variable along the predicted dataset.
\code{type="mask_agg"} output a single heatmap of mask aggregated values,
\code{type="mask_agg"} provides a facet plot of each mask along the model.
}
\examples{
\donttest{
data("ames", package = "modeldata")
ames_fit <- tabnet_fit(Sale_Price ~. , data=ames, epoch=15)
ames_explain <- tabnet_explain(ames_fit, ames)
# Plot the model mask interpretation heatmap
autoplot(ames_explain)
}
 @export
}
