% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat.R
\name{node_to_df}
\alias{node_to_df}
\title{turn a Node object into predictor and outcome data-frame}
\usage{
node_to_df(x, drop_last_level = TRUE)
}
\arguments{
\item{x}{Node object}

\item{drop_last_level}{TRUE unused}
}
\value{
a named list of x and y, being respectively the predictor data-frame and the outcomes data-frame,
as expected inputs for \code{hardhat::mold()} function.
}
\description{
turn a Node object into predictor and outcome data-frame
}
