% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{resolve_data}
\alias{resolve_data}
\title{Transforms input data into a list of
3 torch_tensors being
$x , $x_na_mask, $y
and
cat_idx the vector of x categorical predictor index
cat_dims the vector of number of levels of each x categorical predictor
input_dim  the number of col in \code{x}
output_dim the \code{ncol(y)} in case of (multi-outcome) regression or
the \code{nlevels(y)} in case of classification or
the vector of \code{nlevels(y)} in case of multi-outcome classification}
\usage{
resolve_data(x, y)
}
\arguments{
\item{x}{a data frame}

\item{y}{a response vector}
}
\description{
Transforms input data into a list of
3 torch_tensors being
$x , $x_na_mask, $y
and
cat_idx the vector of x categorical predictor index
cat_dims the vector of number of levels of each x categorical predictor
input_dim  the number of col in \code{x}
output_dim the \code{ncol(y)} in case of (multi-outcome) regression or
the \code{nlevels(y)} in case of classification or
the vector of \code{nlevels(y)} in case of multi-outcome classification
}
