% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{resolve_data}
\alias{resolve_data}
\title{Transforms input data into tensors}
\usage{
resolve_data(x, y)
}
\arguments{
\item{x}{a data frame}

\item{y}{a response vector}
}
\description{
Transforms input data into tensors
}
