% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials.R
\name{decision_width}
\alias{decision_width}
\alias{attention_width}
\alias{num_steps}
\alias{feature_reusage}
\alias{num_independent}
\alias{num_shared}
\alias{momentum}
\alias{mask_type}
\title{Parameters for the tabnet model}
\usage{
decision_width(range = c(8L, 64L), trans = NULL)

attention_width(range = c(8L, 64L), trans = NULL)

num_steps(range = c(3L, 10L), trans = NULL)

feature_reusage(range = c(1, 2), trans = NULL)

num_independent(range = c(1L, 5L), trans = NULL)

num_shared(range = c(1L, 5L), trans = NULL)

momentum(range = c(0.01, 0.4), trans = NULL)

mask_type(values = c("sparsemax", "entmax"))
}
\arguments{
\item{range}{the default range for the parameter value}

\item{trans}{whether to apply a transformation to the parameter}

\item{values}{possible values for factor parameters

These functions are used with \code{tune} grid functions to generate
candidates.}
}
\value{
A \code{dials} parameter to be used when tuning TabNet models.
}
\description{
Parameters for the tabnet model
}
