% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardhat.R
\name{tabnet_pretrain}
\alias{tabnet_pretrain}
\alias{tabnet_pretrain.default}
\alias{tabnet_pretrain.data.frame}
\title{pretrain Tabnet model}
\usage{
tabnet_pretrain(x, ...)

\method{tabnet_pretrain}{default}(x, ...)

\method{tabnet_pretrain}{data.frame}(x, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
}

The predictor data should be standardized (e.g. centered or scaled).
The model treats categorical predictors internally thus, you don't need to
make any treatment.}

\item{...}{Model hyperparameters. See \code{\link[=tabnet_config]{tabnet_config()}} for a list of
all possible hyperparameters.}
}
\value{
A TabNet model object. It can be used for supervised training.
}
\description{
Pretrain the \href{https://arxiv.org/abs/1908.07442}{TabNet: Attentive Interpretable Tabular Learning} model
}
\section{Threading}{


TabNet uses \code{torch} as it's backend for computation and \code{torch} uses all
available threads by default.

You can control the number of threads used by \code{torch} with:\preformatted{torch::torch_set_num_threads(1)
torch::torch_set_num_interop_threads(1)
}
}

\examples{
if (torch::torch_is_installed()) {
data("ames", package = "modeldata")
ames_predictors <- ames \%>\% select(-Sale_Price)
fit <- tabnet_pretrain(x = ames_predictors, epochs = 1)
}

}
